Introduce PIC mute support, initially for powerpc.
Some I-pipe patches may allow to tell the embedded PICs to mute the
low priority interrupt channels, i.e. all IRQs which are not of any
interest to a heading pipeline domain like Xenomai (typically,
Linux-handled device interrupts).

When available, this feature prevents the real-time activity to be
preempted by non-related interrupt events, even for the short time
needed by the I-pipe to log them.

When this support is available from the I-pipe, the Xenomai scheduler
will request all PICs to mute the low priority interrupts when
entering the Xenomai domain, and conversely unmute them when leaving
it.

NOTE: not all platforms may provide such feature, since it depends on
the characteristics of the embedded PICs (typically, how
straightforward it is to mask or lower the priority of a subset of
interrupt sources), and whether the I-pipe implements this feature.