rtcan: ixxat-pci: fix problems with device probing
This patch fixes problems with probing and registration of the
RTCAN IXXAT PCI driver as reported by Stefan Kisdaroczi and
Sebastian Smolorz. The PCI device id table entry now specifies
the PCI sub-system id avoiding the driver's probe function to
be call unnecessarily (a PLX9050 is used for many other non-can
PCI devices). This also makes the checking of the sub-system id
obsolete.

Signed-off-by: Wolfgang Grandegger <wg@grandegegr.com>