nucleus: introduce buffer descriptors
A buffer descriptor is a simple abstraction dealing with copy
operations to/from memory buffers which may belong to different
address spaces.

To this end, the buffer descriptor library provides a small set of
copy routines which are aware of address space restrictions when
moving data, and a generic container type which can hold a
reference to - or cover - a particular memory area, either present
in kernel space, or in any of the existing user memory contexts.

The goal of the buffer descriptor abstraction is to hide address
space specifics from Xenomai services dealing with memory areas,
allowing them to operate on multiple address spaces seamlessly.