nios2: introduce architecture support
This patch merges the arch-specific bits aimed at supporting the nios2
architecture.

This port is still work in progress, even if it already implements a
fully fledged real-time support for nios2. Programs based on the POSIX
skin won't link properly yet. On the other hand, programs based on
skins which do not rely on ld's wrapper magic (-Wl,--wrap) can be
built.

The pipeline patch merged along with this initial material will
require some Xenomai-specific bits to be defined in your FPGA design,
so that a high resolution timer and a dedicated 64bit clock device are
available to the Xenomai core. A complete procedure to update your
design accordingly will be merged in a near future.

In short, this beast already runs on an Altera DE2 board here, but
stay tuned for loading it on your own device.