Make marker string for faults less ambiguous
The trace mark string for faults uses an address
element to notify about the location of the instruction
that caused the fault. "address" is a bit dubious in
this context because it can mean either faulting address
or IP address. Replacing "address" with "ip" also makes
the message more consistent with Linux' page_fault_entry
trace point.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>