rtdm: Add padding to rtser_config
The current layout of rtser_config is unfortunate as it may result in
different layout depending on the compiler alignment setting. Namely,
rx_timeout may be aligned on 8-byte boundaries in user land while it may
not be aligned in the kernel, or vice versa.

Avoid this ambiguity by adding a reserved padding field. Bump profile
revision number due to ABI breakage.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>