hal: check CPU frequency
In some conditions, while testing Xenomai in VirtualBox, I had the
error that xnarch_init_timeconv() is called from init_32.h with a
frequency of 0. That leads to a division by zero, followed by a system
oops.

Of course that is a bug in the virtualisation and Linux reports a CPU
frequency of 0 in /proc/cpuinfo. However, there's no harm if Xenomai
checks that instead of crashing the whole system.

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>