nucleus: Cosmetic cleanup of lostage_handler
Granted, it's unlikely that we will ever hit the last free slot and will
then actually be preempted by the real-time domain that injects another
request. If we were, the slot may be overwritten too early. Let's clean
this up, at least for the sake of the reader that tries to grab the
logic.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>