nucleus: Fix endless loop of DEBUG_SYNCH_RELAX
When a thread is about to block on a mutex resource that is currently
held by a relaxed thread, it will receive SIGXCPU if the relaxed owner
debugging feature is enable. But after the signal is delivered and
handled, the sleeper will restart the mutex service and then receive the
same signal over and over again.

Fix this fatal loop by temporarily disabling the warning for the sleeper
after the first delivery, but also rearming it for the next run right
after that. This will still cause re-deliveries for the same blocking
call if the sleeper may fall out due to a different signal delivery. But
endless loops will be avoided and the solution is kept simple.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>