nucleus: Try to send SIGXCPU to runaway threads first
Runaway shadow threads that perform syscalls need not be killed
immediately by the watchdog. We can first try to send out SIGXCPU and
trigger a migration on syscall return in order to let normal signal
handling / gdb process the failure. This eases the identification of
overruns and also reduces the risk of inconsistent system states due to
hard terminations.

If the target thread does not react (ie. migrate) within another
watchdog timeout period, we kill it nevertheless.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>