nucleus: Add support for sharing data between Xenomai kernel and userland
A new structure (struct xnshared) is introduced. It allows us to share
generic data between user and kernel of Xenomai; a bitmap of feature
flags located at the beginning of the structure identifies which data
are shared. The structure is allocated in the global semaphore heap,
and xnsysinfo.xnshared_offset identifies the offset of the structure
within the heap.

Currently, no shared features are yet supported - the patch only
introduces the necessary ABI changes. When the need arises
to share data between said entities, the structure must
be accordingly extended, and a new feature bit must be added
to the flags.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>