nucleus: Fix RPI issue.
Fix a deadlock which was occuring with short periods.

It was due to the fact that a Linux thread in the RPI queue, when suspended
whereas not current, could not be removed from the RPI queue because the
nucleus was not notified.

We fix the issue by removing such threads from the RPI queue when they head
the RPI queue.

Furthermore we suppress the removal of such threads which was occuring when
Linux was switching them out, to only remove them lazily when they head the
RPI queue.