analogy: add a4l_sizeof_sudb function
This function is only useful for DIO subdevices. Digital subdevices
are a specific kind of subdevice on which channels are regarded as
bits composing the subdevice's bitfield. During a DIO acquisition,
all bits are sampled. Therefore, a4l_sizeof_chan() is useless in
this case and we have to use a4l_sizeof_subd().
With bitfields which sizes are 8, 16 or 32, there is no problem
finding out the size in memory (1, 2, 4); however with widths like
12 or 24, this function might be helpful to guess the size needed
in RAM for a single acquired element.