nucleus: rework skin reference counting
The previous implementation fell short of actually protecting against
deregistration of busy skins.

Just simplify the whole stuff by counting thread attachments (to API
modules) and process bindings (to the system skin). The
non-operational-yet state that exists between the binding to the
system skin, and the actual attachment to the API skin, is inherently
protected by the kernel returning -ENOSYS to userland, when the API
module is not reachable anymore.