Rename SIGXCPU to SIGDEBUG and deliver reason via siginfo
Now that we raise SIGXCPU not only on unwanted migrations or missing
memory locking but also watchdog events, it's getting fairly hard for
user space to tell those reasons apart.

This patch collects the various reasons and attaches them in a siginfo
value to the renamed SIGDEBUG signal. It even allows to find out the
precise reason for migrations, ie. if a syscall, signal delivery, fault,
or detected priority inversion let the nucleus switch some task to Linux
mode.

The existing example is updated accordingly.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>