testsuite/latency: fix false positive in mode switch detection
When ^C is depressed to stop the test, receiving SIGINT on the
sampling task must have caused a mode switch notification
earlier. This is a false positive wrt involuntary mode switch
reporting.

To avoid this, block SIGINT (and most termination signals altogether)
for test threads, and receive them explicitely on the main thread via
sigwait().