rtdm: tag syscalls as conforming
Most RTDM syscalls should be tagged as both adaptive and conforming,
instead of bearing the __xn_exec_current bit.  Actually, we do want
the current domain to change when it is not the most appropriate,
which __xn_exec_current prevents so far.

__xn_exec_current allows application code to infer/assume that the
RTDM driver might behave differently depending on the current runtime
mode of the calling thread, which is wrong.

A conforming syscall makes the Xenomai threads migrate to primary mode
automatically, before RTDM entry points are invoked, since this is the
preferred mode of operation for them.

sys_rtdm_open, sys_rtdm_socket and sys_rtdm_close are not made
conforming though, since real-time resource allocation of RTDM drivers
is already deprecated, and the related open_rt/close_rt callbacks are
scheduled for removal in Xenomai v3.