RTDM: Add rtdm_rt_capable() service
This adds rtdm_rt_capable(), a function that can be used by drivers to
detect callers that could issue a service request also from the
(typically preferred) real-time context. If that is the case, the driver
can trigger a restart of the request if the current context is not
real-time.

CC: Philippe Gerum <rpm@xenomai.org>
CC: Alexis Berlemont <berlemont.hauw@free.fr>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>