mutex-torture: Add test for disallowed lock stealing
This test case reflects an existing bug in the lock stealing
implementation: A lock is stolen from A and then released again. When A
wakes up, it acquires the lock via the in-kernel fastpath and holds it
while being busy. B preempts A and tries to acquire the lock as well. So
far it steals it due to the bug while it should simply boost A and wait
for the release.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>