nucleus/shadow: fix RPI race in xnshadow_relax() introduced by 695f3d0
This patch fixes a race in the RPI management code which may cause the
root thread to remain priority-boosted indefinitely. The related bug
was introduced in the 2.5 timeframe by commit 695f3d0 (aka killer
optimization).

The faulty code fragment from xnshadow_relax(), is this one:

	rpi_push(thread->sched, thread);
	splhigh(s);
	schedule_linux_call(LO_WAKEUP_REQ, current, 0);

The sequence of events leading to the bug is as follows:

- a Xenomai thread running in primary mode enters xnshadow_relax(), to
  move to the Linux domain; because this thread runs under the control
  of the Xenomai scheduler at this point, it is blocked linux-wise
  (TASK_INTERRUPTIBLE).

- a real-time timer tick preempts the thread and switches it out,
  right after it linked itself to the local RPI queue calling
  rpi_push(), but before it had a chance to send an asynchronous wake
  up request for itself to the linux kernel
  (i.e. schedule_linux_call(LO_WAKEUP_REQ...).

As a consequence of this, a blocked shadow linux-wise was introduced
into the RPI queue, with no later opportunity to leave it,
indefinitely preventing the root thread to be demoted to its regular
idle priority.

That bug is easily reproduced by running the latency test concurrently
to the mutex-torture-native test.

The fix consists of making the two operations described above, belong
to a single atomic sequence, by proper interrupt masking.