analogy: make the command registering perform in NRT context
Theorically, the command ioctl could be performed in
real-time. However, with the introduction of rtdm_rt_capable(), we
were able to notice a problem in the NI driver: some
pci_alloc_consistent() calls could be done inside some command
callbacks. So, the quickest fix was to confine cmd ioctls to NRT
domain. The influence is minimal as the next syscall (insn trigger or
read / write) will make the process switch.