rtai: mark as deprecated
Keeping this emulator does not make much sense anymore. Most people
who ported to Xenomai already moved to either the native or POSIX
APIs.

Additionally, this kernel-only emulator has no future in the next
Xenomai architecture, and it is very unlikely that anyone would bother
writing an LXRT emulator for Xenomai.

The RTAI emulator is scheduled for removal in Xenomai 2.6.x.