powerpc: introduce HW_MATH_EMU symbol
We need a way to identify platforms for which kernel-based FPU support
is not available, because some form of math emulation is required to
handle FP code.

Introduce and check CONFIG_XENO_HW_MATH_EMU for this purpose; that
conditional symbol is set by the arch-dep Kconfig frag, depending on
the presence of other platform-specific symbols it may test.