nucleus: Infrastructure for CLOCK_HOST_REALTIME
Augment the shared vdso page with all data required to
implement a clock CLOCK_HOST_REALTIME that provides time
information synchronised with the NTP-corrected time
in the Linux kernel. Data exchange between Linux and
Xenomai is handled via an ipipe event.

Additionally, the patch creates an instance of struct xnvdso
independently of pervasive rt support. This avoids repeating
the definitions for hostrt in both cases.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>