nucleus: demote RPI boost upon linux-originated signal
This patch clears the RPI boost for any relaxed thread which is
getting a signal while running, so that it won't keep a boosted
priority to process asynchronous linux-originated events, such as
termination signals.

Rationale: RPI is mainly for preventing priority inversion during
normal operations in secondary mode, handling signals should not apply
there, since this would also boost low-priority cleanup work, which is
unwanted. The thread may get RPI-boosted again the next time it
resumes for suspension, linux-wise (if ever it does).