nucleus: requeue blocked non-periodic timers properly
Single-stepping into a Xenomai application should freeze timers to
avoid overruns, until the program is continued. Unfortunately,
non-periodic timers are not requeued when the time base is unblocked,
preventing their timeout handler to be fired, thus causing the tasks
pending on them to hang indefinitely.

This patch requeues those timers properly, by interval of 250 ms,
until the timebase is unlocked.