fix private heap unmapping upon fork.
In order to fix private heap unmapping/remapping behaviour with dlopen, an
atfork handler was installed which unmaps and remaps the private heap by
commit e70ce487ac7ab62cd8de28af8ddccf7309f1259d

Unfortunately, the atfork handler for semaphore heaps used a system call to
get private heap data which turned out returning data for the shared heap
after fork, leading to:
- a simple segfault if the private heap and shared heap have different
sizes;
- silent corruption of the shared heap when using private objects in the
child otherwise.

So, we fix this by only unmapping the private heap upon fork, and waiting
for any skin to be bound in the child process for remapping the private
heap.