hal/generic: inline APC scheduling code
rthal_apc_schedule() may be called on the hot path, its implementation
is simple and we need it to be as fast as possible. This patch inlines
it.

In the same move, error checking on the APC parameter is either
removed (APC scheduling) or converted to a BUG_ON() assertion, since
that interface is way too internal for recovering sanely from such a
misuse.