nucleus/shadow: shorten the uninterruptible path to secondary mode
Switching a thread from primary to secondary mode entails running a
significantly long code path with interrupts off, to hand over the
relaxing thread to the Linux scheduler.

Investigation on different architectures showed that such code path
was involved most of the time in latency peaks, typically when an
interrupt arrives at the very beginning of the migration sequence, and
remains blocked until the thread is fully switched out. Having RPI
enabled may increase the penalty, since pushing the relaxing thread to
the local RPI queue is part of this sequence (rpi_push).

Tracing reveals that a significant portion of the uninterruptible
sequence is actually spent running the rescheduling procedure
(xnpod_schedule). However, nothing requires us to suspend /and/ switch
out a relaxing thread atomically; actually, this is even inefficient,
since this tends to give a high priority to a thread going for less
real-time guarantees, over a real-time activity which could be started
by a pending interrupt.

This patch introduces a special handling of the XNRELAX bit condition
in xnpod_suspend_thread(), so that all locks (smp and local
interrupts) are dropped right before switching out the current thread,
to open a window for interrupt preemption. Additionally, interrupt
management is now shared between xnshadow_relax() and
xnpod_suspend_thread(), so that basic assumptions can be made on the
current interrupt state, to further reduce interrupt masking.

Best cases:

- no interrupt will be pending, so the relaxed thread will be switched
  out immediately.

- an interrupt will be pending for the Xenomai domain, performing
  time-critical duties such as waking up a real-time thread, in which
  case the latency to handle a real-time event will have been lower.

Worst case:

- an interrupt will be pending for Linux, in which case the
  rescheduling will be postponed until the interrupt pipeline has
  logged it (but not dispatched, since we will be running over the
  high priority Xenomai domain).