nucleus/sched: prevent remote wakeup from triggering a debug assertion
Jan:
----
"The task that was scheduled in without XNRESCHED set locally has been
woken up by a remote CPU. The waker requeued the task and set the
resched condition for itself and in the resched proxy mask for the
remote CPU. But there is at least one place in the Xenomai code where
we drop the nklock between xnsched_set_resched and xnpod_schedule:
do_taskexit_event (I bet there are even more). Now the resched target
CPU runs into a timer handler, issues xnpod_schedule unconditionally,
and happens to find the woken-up task before it is actually informed
via an IPI."

Gilles:
-------
"Yes, and whether we set the bit and call xnpod_schedule atomically
does not really matter either: the IPI takes time to propagate, and
since xnarch_send_ipi does not wait for the IPI to have been received
on the remote CPU, there is no guarantee that xnpod_schedule could not
have been called in the mean time.

More importantly, since in order to do an action on a remote
xnsched_t, we need to hold the nklock, is there any point in not
setting the XNRESCHED bit on that distant structure, at the same time
as when we set the cpu bit on the local sched structure mask and send
the IPI? This way, setting the XNRESCHED bit in the IPI handler would
no longer be necessary, and we would avoid the race."

What this patch does is exactly that, in an attempt to make the remote
rescheduling code safer and simpler:

- by testing XNSCHED in __xnpod_test_resched() instead of the resched
  bitmask for the current CPU; this bitmask is now only used to
  broadcast the IPI to the CPUs pending a reschedule, from the local
  processor POV.

- by setting the XNSCHED bit immediately in the remote scheduler's
  status, which fixes the unwanted assertion.

See there for the discussion regarding this issue:
https://mail.gna.org/public/xenomai-core/2010-08/msg00084.html