native: fix semantics of broken T_PRIMARY with T_CONFORMING
T_PRIMARY was often misused, leading to applications forcing their
execution domain for no reason most of the time (typically because
Xenomai syscalls do migrate the caller to the proper domain when
needed). As a matter of fact, the only meaningful usage of such
feature is to force a real-time shadow back to primary mode, before
entering a syscall-less loop.

Additionally, now that auto-relax is enabled for non real-time shadow
threads upon syscall exit, forcing T_PRIMARY on such threads would
make no sense at all anymore.

Therefore, to reduce the odds of misinterpreting what the feature
brings, this patch removes the T_PRIMARY bit, and replaces it by
T_CONFORMING, to request that the caller be moved to its preferred
execution domain if it is not yet current.

This results in the following changes API-wise:

 - rt_task_set_mode(clrmask, setmask=T_PRIMARY, &oldmode)
   becomes:
   rt_task_set_mode(clrmask, setmask=T_CONFORMING, &oldmode), forcing
   primary mode for SCHED_FIFO Xenomai threads only. Nop otherwise
   (i.e. a non real-time shadow remains in relaxed mode after the
   syscall has returned, since this is its conforming execution mode).

 - rt_task_set_mode(clrmask=T_CONFORMING, setmask, &oldmode)
   always begets -EINVAL.