src: introduce the Copperplate interface
In order to move RTOS emulators to user-space, we need the rough
equivalent of the nucleus building blocks, available on top of the
POSIX threads layer.

This new interface has to be "ductile" and flexible enough to
implement the various kinds of window dressings we want for the RTOS
APIs, but also to accomodate both the dual kernel and single kernel
environments.

To this end, the patch introduces the Copperplate library, which
abstracts the real-time resources we will need for mimicking RTOS
APIs.

The POSIX threads support it is based on shall be provided by the
Cobalt core in dual kernel mode via the POSIX skin, or directly by the
nptl for the Mercury core, when running over a single kernel image
system (typically, PREEMPT_RT).