posix: introduce SIGSUSP/SIGRESM for pthread_kill()
In order to support the red core layer, we need a way to
suspend/resume a thread synchronously, using the low-level nucleus
services.

To this end, this patch introduces two pseudo-signals, namely SIGSUSP
and SIGRESM, to respectively call xnpod_suspend_thread() or
xnpod_resume_thread() for the target thread.

CAVEAT: those are pseudo-signals above SIGRTMAX, so they can't be
blocked, or queued, or even set in a sigset. They are just meant for
internal usage. They are semantically very different from POSIX's
SIGSTOP/SIGCONT, since:

- they are thread-specific

- they are guaranteed to perform the action synchronously, i.e. the
  target thread shall be suspended or resumed before the call
  returns. If SIGSUSP is sent to the current thread, pthread_kill()
  only returns after SIGRESM is received conversely, sent from another
  thread.