nucleus/sched: Fix fallout of f6af9b831
By re-enabling interrupts in the middle of xnpod_suspend_thread(), f6af9b831
allows an interrupt to happen when running __xnpod_schedule() as a result of
xnshadow_relax(), right between the time xnarch_escalate() is called, and the
time the interrupts are shut off (and the nklock locked).

When this happens, __xnpod_schedule() critical section is run while the current
Adeos domain is Linux, which is precisely what xnarch_escalate() was called to
avoid.

So, we simply shut interrupts off in xnpod_suspend_thread() right after
enabling them for the XNRELAX case, since this is the only place where this
issue may happen.