fastsync: explain invalid calls to xnheap_free
The check mutex->cpid == current->pid in rt_mutex_delete is too eager: it
will prevent mutexes from being rightfully destroyed when the ppd cleanup
callback is invoked, as in this case, the current process is not necessarily
the one which created the mutex.

The problem this check was trying to avoid was to call xnheap_free with
destroyed/wrong heap, or fastlock. But we know that xnheap_free is sufficiently well behaved for such a call to be harmless. So, clearly explain it in a
commentary. At this chance, fix posix skin commentary for the same case,
which no longer matched the code since commit
cb0d865bac11c0090f6f0e3b84e72590cde5e674.