nucleus/shadow: require lostage for __xn_sys_sem_heap
We may get into a page fault due to ondemand mapping when copying the
heap descriptor back to userland in xnshadow_sys_sem_heap(), prior to
relaying faults to the nucleus, since the calling thread is not yet
mapped at this stage, therefore the CPU exceptions are not routed
through the nucleus yet for this thread.

In this case, leaving the exec bits for __xn_sys_sem_heap syscall as
__xn_exec_any would cause the Linux page fault handler to run over the
Xenomai domain, with no interception from the nucleus to downgrade the
stage to the Linux one.

We tag __xn_sys_sem_heap with the __xn_exec_lostage exec bits to
prevent this.