native: make MPS fastlock-compatible
This patch fixes a long-standing bug plaguing the message passing
services since the introduction of the fast locking support
(CONFIG_XENO_FASTSYNCH).

To this end, the per-task client/server synchronization object is
specifically allocated and initialized in the fast sync case, to set
the initial and permanent ownership the right way. This also fixes an
incompatibility with the recent auto-relax feature for non real-time
shadow threads.