nucleus: fix ppd cleanup
During ppd cleanup, current->mm is not necessarily the mm of the process
being cleaned up, this prevents xnshadow_ppd_get from being used during
cleanup. Additionally, current->mm is NULL when executing shadow cleanup
hooks during process exit.

This commit uses an adeos ptd in order to store the mm of a shadow. This
ptd is set upon shadow map to avoid the second issue, and temporarily to
the process executing the mm cleanup to avoid the first one.