nucleus: Drop panic message buffering
The kernel only takes up to 1K per printk, so buffering up to 16K is
already pointless. If we happen to write out more (or enter an endless
loop), we trigger a secondary WARN_ON in vsnprintf first. Finally, and
that's worst, if we crash or lock up while filling nkmsgbuf, nothing is
printed at all.

So let's drop this buffering and write out the panic message line-wise.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>