nucleus/shadow: wakeup the gatekeeper immediately when possible
223685ce enabled the task hardening code over hybrid PREEMPT_RT +
I-pipe enabled kernels, by delaying the wake up call for the
gatekeeper to the schedule event handler.

We may delay the wake up call over non-RT 2.6 kernels as well (and
actually did for a while), but we may not do this when running over
2.4 (scheduler innards would not allow this, causing weirdnesses when
hardening tasks). So we always do the early wake up when running
non-RT, which implicitly includes 2.4 kernels.