nucleus: Fix race between gatekeeper and premature task resumption
A task that is about to be hardened may receive a signal. In that case
the schedule() that the task invoke will set its state to TASK_RUNNING
again and may than decide to move it to a different CPU as also the
gatekeeper is runnable at that point. Now we start a race between the
task continuing to run in Linux context or terminate (on SIGKILL etc.)
and the gatekeeper still holding a reference to this object and
potentially seeing the task state as TASK_INTERRUPTIBLE again (if it
waits long enough).

Avoid this race by waiting for the gatekeeper to drop the obsolete
migration request before returning from xnshadow_harden. Based on
suggestion by Gilles.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>