posix: Do not prevent double-inclusion of original stdio.h
iostream, eg., pulls in stdio.h for internal purposes. This pull will
only define a subset of what an explicit stdio.h inclusion provides. But
as we block a second inclusion so far, this full definition set cannot
be pulled later on. So move #include_next out of the double-inclusion
check.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>