posix: rename as the cobalt interface
include/posix contents have moved to include/cobalt, lib/posix is
renamed lib/cobalt and libpthread_rt becomes libcobalt. Conversely,
the nucleus core becomes a building part of the cobalt interface.

Rationale: the Cobalt-based POSIX interface becomes the core API over
which all copperplate-based APIs are based. It is therefore defining
for this real-time core, and as such should inherit its name.

This also removes any occurrence of posix/, which makes the naming
less ambiguous with respect to glibc's POSIX support we otherwise rely
on fully when based over the Mercury core.

This change conveys a strong hint about the fact that any additional
skin we may want to introduce should be built over the copperplate
interface in userland, and not directly implemented over the
nucleus. This is a requirement for allowing its use over the Mercury
core seamlessly. To this end, both the Cobalt/POSIX skin and the
copperplate library can be extended internally to support the required
features (e.g. additional scheduling policies).