Kconfig: reorganize and simplify, build POSIX, RTDM unconditionally
Rationale: Without the Cobalt/POSIX API, no copperplate-based API
could ever run, so we want it built in. Enabling RTDM unconditionally
removes ugly dependencies in the code (see former select() support),
makes connecting to built-in internal devices an option to replace
nucleus-specific syscalls in the future.

Additionally, RTDM only amounts to 21Kb on powerpc 6xx anyway,
compared to 89Kb for the nucleus and even more for the POSIX
interface.