copperplate: introduce CLOCK_COPPERPLATE
We define the Copperplate clock as a monotonic, non-adjustable
one. This means that delays and timeouts won't be affected when the
kernel host date is changed.

The implementation provides support for absolute dates internally,
with a per-clock epoch value, so that different emulators can have
different system dates.

This also has the advantage of allowing for a single monotonic time
source for both TSC and date (i.e. ns-based) values when running over
the Mercury core, via a unique clock_gettime() call.