copperplate: introduce error backtracing support in debug mode
This patch introduces the __bt() macro, which should be used when
error returns are sent back to function callers, to log the event into
the current per-thread call backtrace. Dumping such backtrace is a
convenient way to determine the exact error path taken by a failing
code.

The other option would be using GDB and the "bt" command, but this
falls short of pointing at the inner error spot, and may also slightly
alter the runtime behavior of applications, sometimes papering over
the issue, particularly over the Cobalt core (but not only).

We also introduce the error_hook() weak routine, which is called when
the inner error notification is received. Used in conjunction with
GDB, this make possible to breakpoint at the original failing point
from a backtrace (i.e. inner call to __bt() reporting an error) for
looking at the issue.

This feature is enabled by --enable-debug.