copperplate/threadobj: cache thread priority locally
We just cannot afford calling the Cobalt core for retrieving the
current thread priority, so we just cache it into the thread object
descriptor.

Granted this might cause mismatches in case this priority is changed
by an external program NOT using the copperplate interface, but doing
so might also break the queuing order in synchronization queues,
regardless of caching.

For consistency reason, we change threadobj_get_priority() to return
the copperplate-cached value for both cores, even if the glibc has its
own cache when calling pthread_get_schedparam() over Mercury.