copperplate/threadobj: guarantee innocuous thread state to suspend hook
We guarantee that a THREADOBJ_SUSPEND event is sent only once the
target thread is in an innocuous state, i.e. about to suspend if
current, or suspended otherwise. This way, the hook routine may always
safely assume that the thread state in userland will not change, until
that thread is resumed.

Conversely, we guarantee that a THREADOBJ_RESUME event is sent while
the target thread is still in an innocuous state, prior to being
actually resuled. This way, the hook routine may always safely assume
that the thread state in userland will not change, until that point.

Fix the Cobalt implementation of threadobj_suspend/resume()
accordingly.