cobalt: introduce extended semaphore ops
We introduce sem_init_np() and sem_broadcast_np(), respectively for
creating a POSIX semaphore which allows for extended control, and for
broadcasting a signal to all waiters of a semaphore.

sem_init_np() adds support for "pulse" mode (i.e. semaphore used as a
discrete event object), FIFO ordering of the wait queue, and extended
report from sem_getvalue().