copperplate: introduce semaphore object (semobj)
Emulating a semaphore via syncobj only makes sense if the extended
operations we may need are not readily available from the underlying
real-time core. These operations are specifically :

- broadcasting a semaphore to unblock all waiters in a single
  operation as if all wait requests had been satisfied.
- "pulse" mode behavior, making the semaphore usable as a discrete
  event object.
- POSIX.1-2001 allowance for having sem_getvalue() return a negative
  number whose absolute value is the count of the number of threads
  currently blocked in sem_wait(3).

Since b97dcdf66, it turns out that we do have those operations
available from the Cobalt interface, directly implemented into the
core system. For performance reasons, we want to use these over
Cobalt, instead of emulating such operations like we still have to do
over Mercury (because they are bluntly non-standard, unavailable from
the glibc).

To this end, this patch introduces the semobj abstraction, which
defines a semaphore object with all the operations the upper
interfaces may need. The semobj package in turn uses the optimal
implementation for the real-time core it is built for.