copperplate/syncobj: introduce syncobj_uninit()
This call is aimed at releasing resources of a syncobj which is
guaranteed to be idle, i.e. no thread could wait on it at the time of
the deletion.

This is specifically aimed at rolling back the effect of
syncobj_init() in a constructor routine, when it fails fully creating
some real-time object. In that case, we don't want to bother with
acquiring a lock on an idle syncobj only for the purpose of deleting
it.