copperplate/heapobj, psos, alchemy: introduce heap inquiry service
The former inquiry service from heapobj was a misnomer, this is
actually a pointer validation routine, which returns the size of the
memory object as a side-effect. We rename it to heapobj_validate().

heapobj_inquiry() is then reused to provide stats for the target
heap. Currently, it only returns its allocated/active size.