cobalt: generalize support for extended scheduling attributes
We provide the necessary bits to create threads from userland within
Cobalt-specific scheduling classes, like SCHED_COBALT, SCHED_SPORADIC
or SCHED_TP.

To this end, we introduce pthread_create_ex(), which accepts an
extended thread attribute structure to hold the class-specific
parameters. In addition, a set of attribute accessors for this
structure is also provided, as a collection of pthread_attr_*_ex()
calls.

pthread_create() is internally rebased on pthread_create_ex(), and the
kernel support for creating Cobalt shadows is changed to accept the
extended attribute set.