cobalt/kernel: remove support for modular build
Rationale: Xenomai's Cobalt in-kernel footprints drastically shrunk to
only include the nucleus, the POSIX/cobalt and RTDM skins. On armv6
for instance, this typically amounts to less than 150Kb, most of it
from the data/bss sections. In the past, most people tended to build
Xenomai 2.x statically into the kernel already, simply because:

- dealing with modules during system updates is error-prone unless
  versioning is enabled, which is not particularly space-efficient.

- matching virtual addresses from kernel message logs to code from
  disassembled vmlinux images is a no-brainer (e.g. during the debug
  phase, after a BUG() assertion triggered).

- enabling the module machinery in the kernel consumes a significant
  amount of RAM. So space-conscious setups will likely avoid them in
  any case.

On the other hand, modules do not bring any real upside. Typically,
the perceived ability to reinit an application system by virtue of
unloading modules then pushing them again is a fairly red herring: if
something went bad in the module space, then the whole kernel is
likely to have been affected, so this operation is pointless anyway.

This patch removes the ability to build the Xenomai core in a modular
fashion, only retaining the static build for the nucleus-cobalt-RTDM
triplet which constitutes the Cobalt core. However, RTDM drivers can
still be built as modules though.